# Adjust model specs for faster CI runs