… spec/models/course/scopes_spec.rb
… spec/models/course_spec.rb
… spec/models/courses/assign_program_type_spec.rb