## Context

Our CI distributes test files to test runners in order to balance the time it takes to run the full suite.

One of the files takes much longer than the others. This PR splits the test examples from that file into other files so the test suite can be better balanced.

I started thinking it was the course_spec taking a long time, but it turned out to be a few tests in the provider_spec.
I've left the changes in the course_spec in the PR as they are not a bad idea IMO.

## Changes proposed in this pull request

Split up the `course_spec.rb`
stop using `is_expected` in `provider_spec.rb`, for some unknown reason each test took ~55 seconds to run.
I've combined 3 of the assertions into one example and they are much quicker.



|Before|After|
|---|---|
|4m23s|1m16s|
|![image](https://github.com/user-attachments/assets/2412d23a-4c13-4571-81a2-bf19bbbfb7a9)|![image](https://github.com/user-attachments/assets/21308f22-bf02-4875-93bf-1e7c5c48600c)|

## Guidance to review

<!-- How could someone else check this work? Which parts do you want more feedback on? -->

## Trello 
https://trello.com/c/UKVVrOOU/684-split-up-coursespecrb-for-faster-ci-runs

## Checklist

- [ ] I have moved hard-coded strings to locale files.
- [ ] I have removed the usage of `data-qa` attributes in HTML files and updated the corresponding tests.